Move back to mainwindow
