+++ src/vorta/views/utils.py
@@ -1,10 +1,5 @@
-from PyQt5 import QtCore
-from PyQt5.QtWidgets import QMessageBox
 from PyQt5.QtGui import QIcon, QImage, QPixmap
-
-from vorta.i18n import translate
 from vorta.utils import uses_dark_mode, get_asset
-from vorta.models import BackupProfileModel
 
 
 def get_colored_icon(icon_name):
@@ -18,33 +13,3 @@ def get_colored_icon(icon_name):
     svg_img = QImage.fromData(svg_str).scaledToHeight(128)
 
     return QIcon(QPixmap(svg_img))
-
-
-def process_log(self, context):
-    cmd = context.get('cmd')
-    if cmd is not None and cmd != 'init':
-        msgid = context.get('msgid')
-        repo_url = context.get('repo_url')
-        if msgid == 'LockTimeout':
-            profile = BackupProfileModel.get(name=context['profile_name'])
-            msg = QMessageBox()
-            msg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
-            msg.setParent(self, QtCore.Qt.Sheet)
-            msg.setText(
-                translate(
-                    "MainWindow QMessagebox",
-                    f"The repository at {repo_url} might be in use by another computer. Override it and continue?"))
-            msg.button(QMessageBox.Yes).clicked.connect(lambda: self.break_lock(profile))
-            msg.setWindowTitle(translate("MainWindow QMessagebox", "Repository In Use"))
-            self._msg = msg
-            msg.show()
-        elif msgid == 'LockFailed':
-            msg = QMessageBox()
-            msg.setParent(self, QtCore.Qt.Sheet)
-            msg.setText(
-                translate(
-                    "MainWindow QMessagebox",
-                    f"You do not have permission to access the repository at {repo_url}. Gain access and try again."))
-            msg.setWindowTitle(translate("MainWindow QMessagebox", "No Repository Permissions"))
-            self._msg = msg
-            msg.show()
